Refactor quicknetsand remove duplication
